Moved share modal
