+++ public/app/features/dashboard/components/ShareModal/ShareModalCtrl.ts
