+++ public/app/features/dashboard/components/ShareModal/index.ts
@@ -1,0 +1,1 @@
+export { ShareModalCtrl } from './ShareModalCtrl';
